+++ index.js
@@ -31,6 +31,7 @@ function cli(process, beforeMinifyCallback) {
     .option('-f, --format <options>', 'Controls output formatting, see examples below')
     .option('-o, --output [output-file]', 'Use [output-file] as output instead of STDOUT')
     .option('-O <n> [optimizations]', 'Turn on level <n> optimizations; optionally accepts a list of fine-grained options, defaults to `1`, see examples below, IMPORTANT: the prefix is O (a capital o letter), NOT a 0 (zero, a number)', function (val) { return Math.abs(parseInt(val)); })
+    .option('--batch-suffix <suffix>', 'A suffix (without extension) appended to input file name when processing in batch mode (`-min` is the default)', '-min')
     .option('--inline [rules]', 'Enables inlining for listed sources (defaults to `local`)')
     .option('--inline-timeout [seconds]', 'Per connection timeout when fetching remote stylesheets (defaults to 5 seconds)', parseFloat)
     .option('--remove-inlined-files', 'Remove files inlined in <source-file ...> or via `@import` statements')
@@ -172,6 +173,7 @@ function cli(process, beforeMinifyCallback) {
   }
 
   var configurations = {
+    batchSuffix: inputOptions.batchSuffix,
     beforeMinifyCallback: beforeMinifyCallback,
     debugMode: debugMode,
     removeInlinedFiles: removeInlinedFiles,
@@ -238,7 +240,7 @@ function minify(process, options, configurations, data) {
 
     if (options.batch && !('styles' in minified)) {
       for (inputPath in minified) {
-        processMinified(process, configurations, minified[inputPath], inputPath, toOutputPath(inputPath));
+        processMinified(process, configurations, minified[inputPath], inputPath, toOutputPath(inputPath, configurations.batchSuffix));
       }
     } else {
       processMinified(process, configurations, minified, null, options.output);
@@ -246,8 +248,10 @@ function minify(process, options, configurations, data) {
   });
 }
 
-function toOutputPath(inputPath) {
-  return inputPath.replace(/\.css$/, '-min.css');
+function toOutputPath(inputPath, batchSuffix) {
+  var extensionName = path.extname(inputPath);
+
+  return inputPath.replace(new RegExp(extensionName + '$'), batchSuffix + extensionName);
 }
 
 function processMinified(process, configurations, minified, inputPath, outputPath) {